extract cluster_manager to separate module
